Expense categories now included in same page, as well as messages to user when submitting
